split out character sheet rendering & fill empty space on situation and cmd & add drawing box points calculator utility function and use that for drawing
