+++ screen/cmd.go
@@ -230,17 +230,21 @@ func (screen *GameScreen) renderUserCommands() {
 	infoLines = infoLines.appendRefreshCmd(screen)
 	infoLines = infoLines.appendEndGameCmd(screen)
 
-	// box start point (x, y)
+	// cmd box start point (x, y)
 	x := 2
-	y := screen.screenSize.Height/2 + 1
+	y := screen.cmdInnerStartY()
+	w := screen.leftInnerWidth()
+	h := screen.cmdInnerHeight()
 
 	bgcolor := uint64(bgcolor)
 	fmtFunc := screen.colorFunc(fmt.Sprintf("255:%v", bgcolor))
 	for index, line := range infoLines {
 		io.WriteString(os.Stdout, fmt.Sprintf("%s%s",
-			cursor.MoveTo(y+index, x), fmtFunc(line)))
-		if index+2 > int(screen.screenSize.Height) {
+			cursor.MoveTo(y+index, x), fmtFunc(fillRightWithSpace(line, w))))
+		if index+2 > int(screen.Height()) {
 			break
 		}
 	}
+
+	screen.drawFill(x, y+len(infoLines), w, h-len(infoLines))
 }
